Remove Canary stuff for now
Epiphany Canary has been dead for about half a year now since it depends
on the WebKit runtime, which does not provide a GTK 4 build of
WebKitGTK. It's confusing to keep unused files around indefinitely so
time to go. Canary is welcome back in the future when it works again.

Part-of: <https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/1319>